Refactor: sprint1 advanced mission 1번째 코드리뷰 반영, TimeHelper로 함수기능 분리
